Move run_prover and run_prover_cli into a dedicated cli module
These two functions are thin process-IO wrappers — they lock real
stdin/stdout and run clap argument parsing — so they cannot be driven
from unit tests without trampling the test runner's own streams. Split
them into crates/dp-zk-cli/src/cli.rs and exclude that file (along with
the bin shims) from SonarCloud coverage analysis. The well-tested
prove_from_bytes and run_prover_io core stays in lib.rs and remains
the only thing whose coverage the quality gate measures.
