+++ crates/dp-zk-cli/src/lib.rs
@@ -6,7 +6,7 @@ use std::process::ExitCode;
 
 use ark_std::rand::rngs::StdRng;
 use ark_std::rand::SeedableRng;
-use clap::{CommandFactory, FromArgMatches, Parser};
+use clap::Parser;
 use dp_zk::witness::BatchWitness;
 use rust_decimal::Decimal;
 use serde::Deserialize;
@@ -214,32 +214,8 @@ pub fn run_prover_io<R: Read, W: Write>(
     }
 }
 
-/// Stdin JSON → stdout proof bytes. Thin process-IO wrapper over
-/// [`run_prover_io`].
-///
-/// Exit codes:
-/// - 0: proof written to stdout.
-/// - 2: stdin read failure, missing private_witness, or invalid input.
-/// - 3: keys missing / version mismatch.
-/// - 4: circuit build, prover, or stdout write failure.
-pub fn run_prover(batch_size: usize, keys_dir: Option<PathBuf>) -> ExitCode {
-    let dir = resolve_keys_dir(keys_dir);
-    run_prover_io(
-        std::io::stdin().lock(),
-        std::io::stdout().lock(),
-        batch_size,
-        &dir,
-    )
-}
-
-/// Parse [`ProverArgs`] under a per-binary program name/about string, then
-/// dispatch to [`run_prover`]. The shared entrypoint for the `dp-zk-cli` and
-/// `dp-aggregator` binaries.
-pub fn run_prover_cli(name: &'static str, about: &'static str) -> ExitCode {
-    let cmd = ProverArgs::command().name(name).about(about);
-    let args = ProverArgs::from_arg_matches(&cmd.get_matches()).unwrap_or_else(|e| e.exit());
-    run_prover(args.batch_size, args.proving_key)
-}
+pub mod cli;
+pub use cli::{run_prover, run_prover_cli};
 
 #[cfg(test)]
 mod tests {
