+++ sonar-project.properties
@@ -10,6 +10,12 @@ sonar.exclusions=**/target/**,contracts/**,**/proto/**,**/build.rs
 
 sonar.coverageReportPaths=coverage/sonar-coverage.xml
 
+# Trivial process-IO shims — `fn main() { dp_zk_cli::run_prover_cli(...) }`
+# plus the wrappers that lock real stdin/stdout/clap. No logic to test in
+# isolation; the prove pipeline they wrap is covered via run_prover_io and
+# prove_from_bytes tests in lib.rs.
+sonar.coverage.exclusions=crates/dp-zk-cli/src/bin/**,crates/dp-zk-cli/src/cli.rs
+
 sonar.cpd.minimumTokens=50
 sonar.cpd.minimumLines=10
